backend/requirements.txt
Django==4.2.3
djangorestframework==3.14.0
black==23.11.0
flake8==6.1.0
isort==5.12.0
django-environ==0.11.2
psycopg2==2.9.6
social-auth-app-django==5.4.1
django-filter==23.2
python-dotenv==1.0.1
django-cors-headers==4.0.0